Estava com conflito no banco de dados por ser long e no banco é int

+++ src/main/java/br/com/project/usuarios/Clientes.java
@@ -6,7 +6,7 @@ import jakarta.persistence.*;
 public class Clientes {
 
     @Id
-    private Long idCli;
+    private int idCli;
 
     @Column(nullable = true)
     private String nome;
